Web/depthMap/requirements.txt
numpy==1.21.2
#onnx==1.10.1
onnxruntime==1.9.0
Pillow==8.3.2